requirements-base.txt
open-dataflow==1.0.6

numpy==1.26.4
datasets==3.2.0
scipy==1.17.0
torch==2.10.0
torchvision==0.25.0
torchaudio==2.10.0
tqdm==4.67.2
transformers==4.57.3
aisuite==0.1.14
math-verify==0.8.0
word2number==1.1
accelerate==1.12.0
rapidfuzz==3.14.3
colorlog==6.10.1
appdirs==1.4.4
datasketch==1.8.0
modelscope==1.33.0
addict==2.4.0
pytest==9.0.2
rich==14.3.2
docstring-parser==0.15
pydantic==2.12.5
nltk==3.9.2
colorama==0.4.6
gradio==6.5.1

# text2sql
func_timeout==4.3.5
sqlglot==28.7.0
PyMySQL==1.1.2

# general text
fasttext-wheel==0.9.2
langkit==0.0.35
openai==1.109.1
sentencepiece==0.2.1
presidio-analyzer==2.2.360
presidio-anonymizer==2.2.360
vendi-score==0.0.3
google-api-core==2.28.1
google-api-python-client==2.187.0
evaluate==0.4.6
contractions==0.1.73
symspellpy==6.9.0
simhash==2.1.2

# knowledge base cleaning
chonkie==1.3.1
trafilatura==2.0.0
lxml_html_clean==0.4.3
pymupdf==1.26.7
httpx==0.27.2

# dataflow agent
cloudpickle==3.1.2
fastapi==0.124.0
pandas==2.3.3
psutil==7.2.2
pyfiglet==1.0.4
PyYAML==6.0.3
requests==2.32.5
termcolor==3.3.0
uvicorn==0.40.0
sseclient-py==1.9.0
supabase==2.27.2

langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
faiss-cpu==1.12.0
onnxruntime==1.23.2
mkdocs-material==9.7.0
beautifulsoup4==4.14.3
playwright==1.56.0
ipython==8.18.1
opencv-python==4.11.0.86

# rebuttal
docling==2.71.0
# websearch knowledge store
playwright-stealth
# speech
librosa==0.11.0
soundfile==0.13.1
python-pptx==1.0.2
scikit-learn==1.8.0